requirements.txt
pymupdf
ollama
Pillow

